ml/surya-ocr/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
fastapi==0.115.2
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fsspec==2024.9.0
ftfy==6.3.0
h11==0.14.0
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
packaging==24.1
pdftext==0.3.16
pillow==10.4.0
protobuf==5.28.2
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
sniffio==1.3.1
starlette==0.40.0
surya-ocr==0.6.3
sympy==1.13.1
tabled-pdf==0.1.0
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
